+++ core/src/lib.rs
@@ -120,6 +120,7 @@ pub mod deno {
 
 pub mod databases_store {
     pub mod gcs;
+    pub mod gcs_background;
     pub mod postgres;
     pub mod store;
 }
